Clean up TimelineWrapper
